gateway: fix fh usage in case of current mode differs to the fh mode
The initial idea was to replace only stale fh to an actual one on any non-open
operation, if the mode was changed.
Unfortunatelly, it doesn't work, becuase libfuse copies fh on each opeartion
call instead of providing the address of the original one (the one, filled in
open() call). This means, that we can replace descriptor, and on each write,
read, etc. call we have to reopen the actual fh and _close_ it. Otherwise fuse
process run out of free fds very fast.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>