gateway: reopen of file handler reintroduced
This time file hadnler reopen works as it should.
Notes:
1) Libfuse stores fh pointer internally and copies it on each upper call. This
means, that fh can't be just replaced on fi object after reopen: on next call
the old pointer will be received again.
The solution is to replace not the pointer, but contents.
2) To reopen a new fh, open flags, stored on old one must be used, because
libfuse garantees to provide correct open flags only on open() and release()
calls.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>