manager: fix mount command handler
The original implementation was very wrong:
1) Remount attempts start before container is ready. Thus -EPERM can be
received on mount. It shouldn't be handled as an error.
2) Remount attempt can get EPROTONOSUPPORT result. This happens when fs module is
not loaded. Also shouldn't be treated as an error.
3) Remount should be done in a loop in proper context (namespaces) with
increasing pause between attempts. Thus, is have to be done in a child
process, joined the context.
4) Mode update after successfull remount have to be done in the parent of
remount process, because it works in initial context, and thus sockets are
accessible for it (child can have mnt and/or user ns joined, and thus not be
able to send a message to a unix socket).

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>