spfs: rewrite logic
This patch is not reviewable. But it's a attempt to fix an clean this initial
mess.
First of all, functions "my_fuse_daemonize" was replaced by "daemon".
Second, all the waiting logic was throwed away as overcomplicated.
Simplier and more conviniet way comes soon.
Third, most of this fuse-not fuse code cleaned and moved to main, which is now
not very simple, but strict.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>