socket: unreliable socket loop introduced
There is a mess with all these socket helpers...
Reliable socket loop send handler result as reply to the caller. But it
doesn't make any sense, if handler forks and doesn't wait for the child.
Moreover, in some cases child shouldn't be waited (like in case of replace
mount), because child can take unpredictable time slot. Or in other way
around: mount command have to be synchronious to make sure, that spfs was
started as a child (spfs is calld via exec in a child, but parent have to make
sure, that spfs started successfully; if parent will be a child, then spfs
becomes a daemon after parent's exit).
So, better send request result not in the main loop, but in the command
handler (forked or not). It just simlier and less error prone.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>