manager: lazily unmount spfs working directory
This appoach with retry doesn't work.
What is even worse, if this directpry and mount point remains in place, next
suspend/resume spfs-manager can reuse it, and it's newly created contents can
be overmounted by tmpfs from the previous run.
In spite of it's still not clear, why mount point is not unmounted when fuse
master has exited (some dangling kernel reference, probably), it's better to
detach the mount fro tree, so it won't be migrated next time.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>