spfs: do not remove xattrs for hidden files on file release
This operation (relase) can be called multiple times (because file belongs to
multiple processes). Thus, releasing attributes on the first call will brake
the logic for the rest of file users.
Attributes will remain until it's unlinked or any other file is renamed into
the hidden one (if any). In this case the will be released in spfs_move_xattrs()

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>