manager: split resources replacement into thre different functions
There are four types of resource: file descriptors, file memory mappings, fs
(cwd and root) and exe link.
In case of multi-threaded application, exe can be changed only when all the
mappings to these file in all threads are unmapped.

For some reason (not clear, unfortunatelly), exec link can't be changed
_before_ fs (cwd, root): fails with -EBUSY.
I failed to find, how this can happen, but the only reasonm is that cwd is
somehow related to mappings.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>