manager: carry spfs reference during replacements
This is required to be able to access extended attributes for silly renamed
file mappings (there is no way to discover, whether mapped file was renamed on
dump or not except accessing xattr, set by criu. To be able to do so, spfs
mount point have to be in hands).
Moreover, this allows to hold spfs mount point, and repeat resources
replacement for whatever reason).
But it's being hold in main spfs-manager process context. And thus has to
be closed to release spfs mount poing upon successful replacement.
This is done in sigchld_handler, if spfs "replacer" process exited with
code 0.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>