swap: change exe link after all other resources have changed
Exe can be swapped only when there are no other links on it (maps).
If mm in shared (in case of threads, for example), all the private mappings
of all the threads have t be unmapped first to be able to change exe link.
So, let's swap all the resources first, and then swap all the exe links.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>