processes: split collect_process_fd
Helper "create_file_obj" is a generic helper, which will be used also for
mapping later.
Helper "collect_fd_obj" is used to open fd and add it to the tree, if a new
one.
Helper "collect_process_fd" now does the thing it was aimed: gets fd and adds
it to process fd list.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>