manager: new file object related helpers introduced
The idea is to use them to collect all the data, required for open an ff, on
file object.
This will allow to open it in any time by just using the reference to the
object.
This, in turn, allows to share fil eobject (and descriptor, if opened) as much
as required.
IOW, by using file object, there is not required to collect open the file
descriptors on examine stage (although it's less safe): fd can be opened
exactly before it's being send to destination process and closed immideatly
after. Thus alllowing to have much less descriptors opened in the same time.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>