manager: detach from process after examine and attach again for swap
Yep... This is bad, but...
CRIU spoling shared FS structure restore. This leads to unique FS struct per
each (!!) thread of a multithreaded process. This, is turn, means, that every
(!!!) thread has to be injected with parasite.
But all the threads sharing the same fd struct, and, at some point, there
won't be enough file descriptors for this.
This already happened with thread-bomb test (1024 threads) on memfd creation.
To overcome this, let's detach from process after examination, and attach and
detach again on swap.

Note: This resources wastage has to be removed has to be removed once CRIU is
fixed.

https://jira.sw.ru/browse/PSBM-50800

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>