manager: revert commit ce578e797811554ae693d9b76fde67326ccbc16c
The mentioned commit was use to overcome SPFS mount problmes, introduced by
recent changes in CRIU, which made it work in VE#0.
But this commit is an overkill (enter and exit from VE#X on each SPFS-related
operation).
Moreover, it introduces an issue: SPFS manager is not killed on restore
failure, because it works in VE#0 (libvzctl kills all the processes in VE#X
cgroup).
So, let's revert ce578e797811554ae693d9b76fde67326ccbc16c.
Fix for the issue will come in the next patch.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>