manager: create new path if "link-remap"'ed exists
Files on NFS are renamed to ".nfs0000XX.XX" upon unlink (if opened).
It's expected, that NFS server will remove them on final close.
Looks like it somethimes doesn't do it (or maybe file is still opened by some
other process somewhere).
Anyway, we have a "link-remap" file, and it's enough to restore the state
(there can be any other dentry, because will be unlinked anyways).

https://jira.sw.ru/browse/PSBM-64299

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>