manager: add "ns_mountpoint" option to mount request
This new parameter is optional and allows to specify two different paths:
1) "mountpoint" -  path to mount SPFS to
2) "ns_mountpoint" - path, where this mout has to be replaced.

In generic case hese paths are the same.
But CRIU needs to distinguish them, because mount points are restore in some
yard first, and then moved/propagated to all the container namespaces.
Thus, when spfs-manager is released by CRIU, there is mountpoint by path
"mountpoint": it was moved into container context.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>