spfs: get work mode instead of copying
This is an optimization and precursor patch.
It optimizes away copying.
Plus it allows to reuse current work mode reliably, thus allowing to pass it
(with proxy directory fd) to proxy operations. Which in turn allows to use
relative to proxy directory fd file operations. Which, in turn, allows to make
Proxy mode independent of current namespace state (including situation, when
proxy directory root was lazy unmounted).

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>