manager: mount spfs in interruptible mode
Using interruptible or uninterruptible mode should depend on target file
system.
Say, it has to borrow this option in case of NFS mount.
But let it be interruptible by default. Otherwise one won't be able to log in,
because bash calls for pidof, which in turn calls for stat on exe links for all
processes in the systems and hungs, if there is one on spfs mount in Stub
mode.
The same applies to NFS mounts, but there it's adviced to mount in
interruptible mode.
So, let's mimic this behaviour.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>