spfs: check system call was interrupted while waiting mode change in Stub mode
This is required to support "interruptible" mode, when system call can be
interrupted by user.
Interruptible mode, in turn, in needed to allow users to log in, when spfs is
in Stub mode, because bash calls for pidof, which, in turn calls stat for in
processes "exe" link in the system, and is any of them in located on spfs -
stuck in it's internals.
This is ok, because exactly this happens with NFS, when server in unreachable.
But NFS in interruptible mode allows to interrupt the request, which allows to
proceed with logic. So SPFS has to do likewise.

In case of "Stub" mode interrupts support is simple: check for
fuse_interrupted once in a while (in 100ms to be precise) and return of EINTR
is required.

https://jira.sw.ru/browse/PSBM-72314

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>