manager: disable readahead for SPFS "restore" mode
CRIU migrates modified pages for private file mappings.
On restore it writes them to the recreated mapping.
The architectural limitation for NFS migration is that files are dummy and
contain garbage. Thus only pages to be replaced have to be read.
But FUSE has readahead enabled by default. Which leads to garbage in mappings
and to segmentation fault shortly.
This patch disables readahead for "restore" mode (special flavour for CRIU)
thus solving the issue.

https://jira.sw.ru/browse/PSBM-73023

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>