spfs: don't copy the first option by default
This is a bogus and strange code.
We need to copy unknown options and mount point. That's all we need.
This, in turn, means, that we need some indicator, whether we need to copy
options (for FUSE daemon) or not: "nind".
It accounts all the processed options (with arguments), and if it's less, than
optind, then we need to copy the rest after options parsing.
And we need to skip argv[0], that's why it's set to optind.
One more thing: we increase nind only once for multiply-allowed options.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>